standalone programs 64-bit clean
git-svn-id: http://opencores.org/ocsvn/eco32/eco32/trunk@259 133310ea-c0b2-4f2d-ae91-eb8ea07367b3

+++ stdalone/dskchk/main.c
@@ -21,7 +21,7 @@
 #define RAND_MAX	0x7FFF
 
 
-static unsigned long randomNumber = 1;
+static unsigned int randomNumber = 1;
 
 
 void srand(int seed) {
